Fix the increaseRgbaAlpha() implementation
Previously there was a hack to only increase the alpha channel to 0.99 to
avoid a bug where the hue appeared to change when alpha reached 1.0.
After consideration, it became clear that the CSS string that was being
returned excluded the alpha value if it was equal to 1.0, which broke the
old parsing logic.